requirements.txt
asgiref==3.5.2
Brotli==1.0.9
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
pymongo==3.12.3
pytz==2022.6
sqlparse==0.2.4
tzdata==2022.6
whitenoise==6.2.0
